+++ src/ItemTreeTraversal/ItemTreeTraversalFactory.php
@@ -12,7 +12,7 @@ use Yiisoft\Db\Connection\ConnectionInterface;
  *
  * @internal
  */
-class ItemTreeTraversalFactory
+final class ItemTreeTraversalFactory
 {
     /**
      * Creates item tree traversal strategy depending on used RDBMS.