+++ tests/Base/ManagerWithDbItemsAndAssignmentsTest.php
@@ -12,22 +12,7 @@ use Yiisoft\Rbac\Tests\Common\ManagerLogicTestTrait;
 
 abstract class ManagerWithDbItemsAndAssignmentsTest extends ManagerTest
 {
-    use ManagerLogicTestTrait {
-        setUp as protected traitSetUp;
-        tearDown as protected traitTearDown;
-    }
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->traitSetUp();
-    }
-
-    protected function tearDown(): void
-    {
-        parent::tearDown();
-        $this->traitTearDown();
-    }
+    use ManagerLogicTestTrait;
 
     protected function createItemsStorage(): ItemsStorageInterface
     {